requirements.txt
openai==0.8.0
elevenlabs==0.1.0
speech_recognition==3.8.1
pyaudio==0.2.11
faiss-cpu==1.7.2
keyboard==0.13.5